requirements.txt
Flask>=2.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
reportlab>=4.0.0
python-docx>=1.1.0
numpy>=1.24.0

# Testing dependencies
selenium>=4.15.0
pytest>=7.4.0
pytest-html>=4.1.0
webdriver-manager>=4.0.0

